small http refactor, fix header() to use proper API
