+++ pubky-client/src/api/mod.rs
@@ -1,6 +1,3 @@
 pub mod auth;
-#[cfg(not(target_arch = "wasm32"))]
-pub mod http_native;
-#[cfg(target_arch = "wasm32")]
-pub mod http_wasm;
+pub mod http;
 pub mod public;
